Initial planet and sun setup for level 4
Note, I tried to add bloom on the sun, I cannot for the life of me figure out how this works. 5 tutorials and 20 forum posts later I have gotten nowhere. I think it might've been because the project was initially setup with SRP and not URP. Even changing requires you to create completely new materials from scratch for the updates to take effect.